chatbot_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
transformers>=4.35.2
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
tokenizers>=0.19.0